requirements.txt
ultralytics>=8.3.0
opencv-python>=4.9.0
roboflow>=1.1.40
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
onnx>=1.16.0
onnxruntime>=1.17.0
